Moved images folder out of routes folder. updated relative paths. renamed two images.
