+++ src/app.css
@@ -10,7 +10,7 @@ body {
     margin: 0; 
     padding: 0;
     color: white;
-    background-image: url('./routes/images/dots3.png');
+    background-image: url('./images/dots3.png');
     font-family: Montserrat, Roboto, Lato, sans-serif ;
     font-weight: 100;
