+++ src/components/LandingImage.svelte
@@ -1,6 +1,6 @@
 <script>
-	import pic1 from '../routes/images/Option5.png';
-	import pic2 from '../routes/images/Option4.png'
+	import pic1 from '../images/man.png';
+	import pic2 from '../images/woman.png'
 </script>
 <div class= "landing-img-container">
 	<img src={pic1} alt="graphic art depicting man using laptop and working on a mobile application."/>
@@ -13,16 +13,9 @@
 		display:flex;
 		width:65%;
 	}
-img{
-	/* width: 400px;
-	border-radius: 2px;
-	min-width: 200px; */
-}
+
 .image-2{
 	transform: scaleX(-1);
 }
-
-/* Q */
-/* how do you make the image responsive (ie - not scrunch up when window size changes?) */
 
 </style>
